flask-jwt-auth/requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-JWT-Extended==3.24.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
passlib==1.7.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.11
Werkzeug==0.16.0
